tweaking final figures and renaming files to represent which figures the script generates

+++ 02_model-vpd-thresholds-using-goes/12_get-thresholds-and-plot-gam.R
@@ -6,38 +6,45 @@ library(patchwork)
 library(sf)
 library(stringr)
 library(here)
+library(data.table)
 
-predictions <- list.files(pattern = "-predictions.csv$", 
-                          path = here("data", "out", "mods"), 
-                          full.names = TRUE) %>%
-  lapply(vroom) %>%
-  bind_rows %>%
-  mutate(lc_name = gsub("_", " ", lc_name))
+if(!file.exists(here::here("data", "out", "gam-predictions.csv"))) {
+  predictions <- list.files(pattern = "-predictions.csv$", 
+                            path = here("data", "out", "mods"), 
+                            full.names = TRUE) %>%
+    lapply(vroom) %>%
+    bind_rows %>%
+    mutate(lc_name = gsub("_", " ", lc_name))
+  
+  # this data frame contains posterior predictive draws for every land cover type
+  predictions
+  
+  data.table::fwrite(x = predictions, file = here::here("data", "out", "gam-predictions.csv"))
+}
 
-# this data frame contains posterior predictive draws for every land cover type
-predictions
-
-
+predictions <- data.table::fread(input = here::here("data", "out", "gam-predictions.csv"))
 
 # Compute vpd thresholds --------------------------------------------------
 # These are posterior means of VPD values for which the probability of a zero
 # (not detecting a fire) is 0.95
 
-thresholds <- predictions %>%
-  mutate(abs_diff = abs(pr_zero - 0.95)) %>%
-  group_by(j, lc_name) %>%
-  filter(abs_diff == min(abs_diff)) %>%
-  ungroup %>%
-  group_by(lc_name) %>%
-  summarize(vpd_thresh_hpa = mean(VPD_hPa), 
-            sd = sd(VPD_hPa)) %>%
-  mutate(probability_of_zero = 0.95) %>%
-  arrange(vpd_thresh_hpa)
+if(!file.exists(here::here("data", "out", "zero-goes-af-vpd-thresholds.csv"))) {
+  thresholds <- predictions %>%
+    mutate(abs_diff = abs(pr_zero - 0.95)) %>%
+    group_by(j, lc_name) %>%
+    filter(abs_diff == min(abs_diff)) %>%
+    ungroup %>%
+    group_by(lc_name) %>%
+    summarize(vpd_thresh_hpa = mean(VPD_hPa), 
+              sd = sd(VPD_hPa)) %>%
+    mutate(probability_of_zero = 0.95) %>%
+    arrange(vpd_thresh_hpa)
+  
+  thresholds %>%
+    write_csv(here("data", "out", "zero-goes-af-vpd-thresholds.csv"))
+}
 
-thresholds %>%
-  write_csv(here("data", "out", "zero-goes-af-vpd-thresholds.csv"))
-
-
+thresholds <- readr::read_csv(here::here("data", "out", "zero-goes-af-vpd-thresholds.csv"))
 
 # Visualize partial effects -----------------------------------------------
 
@@ -59,26 +66,28 @@ plot_df <- partial_df %>%
   left_join(thresholds) %>%
   mutate(lc_name = paste0(lc_name, " (", round(vpd_thresh_hpa, 1), ')'))
 
-partial_plot <- plot_df %>%
-  ggplot(aes(VPD_hPa, mu, color = facet_label)) +
-  geom_path(alpha = .5, aes(group = lc_name)) +
-  geom_ribbon(aes(ymin = lo, ymax = hi, group = lc_name), 
-              color = NA, alpha = .05) +
-  theme_minimal() + 
-  theme(legend.position = "none", 
-        panel.grid.minor = element_blank(), 
-        axis.title.x = element_text(hjust = .24)) + 
-  xlab("Vapor pressure deficit (hPa)") + 
-  ylab("Active fire detection probability") + 
-  facet_wrap(~reorder(facet_label, vpd_thresh_hpa))
-partial_plot
+### Previous iteration, saved for posterity
+# partial_plot <- plot_df %>%
+#   ggplot(aes(VPD_hPa, mu, color = facet_label)) +
+#   geom_path(alpha = .5, aes(group = lc_name)) +
+#   geom_ribbon(aes(ymin = lo, ymax = hi, group = lc_name), 
+#               color = NA, alpha = .05) +
+#   theme_minimal() + 
+#   theme(legend.position = "none", 
+#         panel.grid.minor = element_blank(), 
+#         axis.title.x = element_text(hjust = .24)) + 
+#   xlab("Vapor pressure deficit (hPa)") + 
+#   ylab("Active fire detection probability") + 
+#   facet_wrap(~reorder(facet_label, vpd_thresh_hpa))
+# partial_plot
+# 
+# ggsave(plot = partial_plot, 
+#        filename = here("figs", "vpd-partial-effects.pdf"), 
+#        width = 7.5, height = 3.5)
+# ggsave(plot = partial_plot, 
+#        filename = here("figs", "vpd-partial-effects.png"), 
+#        width = 7.5, height = 3.5)
 
-ggsave(plot = partial_plot, 
-       filename = here("figs", "vpd-partial-effects.pdf"), 
-       width = 7.5, height = 3.5)
-ggsave(plot = partial_plot, 
-       filename = here("figs", "vpd-partial-effects.png"), 
-       width = 7.5, height = 3.5)
 
 # Combine Koppen classes into single facet per Jennifer/Adam request
 partial_plot_single <- 
@@ -87,10 +96,8 @@ partial_plot_single <-
   geom_path(aes(group = lc_name), lwd = 1.0) +
   geom_ribbon(aes(ymin = lo, ymax = hi, group = lc_name), 
               color = NA, alpha = 0.02) +
-  theme_minimal() + 
-  # theme(legend.position = "none", 
-  #       panel.grid.minor = element_blank(), 
-  #       axis.title.x = element_text(hjust = .24)) + 
+  theme_minimal() +
+  theme(text = element_text(family="Helvetica")) +
   labs(x = "Vapor pressure deficit (kPa)",
        y = "Active fire detection probability",
        color = "Köppen-Geiger\nclimate classifications")
@@ -99,10 +106,10 @@ partial_plot_single
 
 ggsave(plot = partial_plot_single, 
        filename = here("figs", "vpd-partial-effects_single-panel.pdf"), 
-       width = 7.5, height = 3.5)
+       width = 183, height = 89, units = "mm")
 ggsave(plot = partial_plot_single, 
        filename = here("figs", "vpd-partial-effects_single-panel.png"), 
-       width = 7.5, height = 3.5)
+       width = 183, height = 89, units = "mm")
 
 
 # Supplementary partial effects plots, with uncertainty ----------------------
@@ -117,11 +124,13 @@ bayes_plot <- predictions %>%
   xlab("Vapor pressure deficit (kPa)") + 
   ylab("Active fire detection probability") + 
   facet_wrap(~fct_reorder(lc_name, vpd_thresh_hpa), 
-             nrow = 5, labeller = label_wrap_gen()) + 
+             nrow = 5, labeller = label_wrap_gen(width = 20)) + 
   geom_path(data = partial_df %>%
               mutate(lc_name = trimws(gsub("\\(.*", "", lc_name))) %>%
               left_join(thresholds), 
             aes(y = mu), size = 1)
+
 bayes_plot
-ggsave(here("figs", "bayes_plot.png"), bayes_plot, width = 9, height = 6)
-ggsave(here("figs", "bayes_plot.pdf"), bayes_plot, width = 5, height = 6)
+
+ggsave(here("figs", "bayes_plot.png"), bayes_plot, width = 183, height = 145, units = "mm")
+ggsave(here("figs", "bayes_plot.pdf"), bayes_plot, width = 183, height = 145, units = "mm")
